Refactored recovery to a separate file
